Move spinner test to timers.test.tsx
Since the test relies on polling, it needs fake timers, otherwise there are timeouts that execute after the test has ended, which messes up other tests.
